+++ lib/storey.rb
@@ -1,22 +1,26 @@
+require 'easy_class_to_instance_method'
+require "active_support/core_ext/module" # so we can use mattr_accessor
+
 require "storey/version"
 require "rails/all"
-require "active_support/core_ext/module" # so we can use mattr_accessor
-require 'easy_class_to_instance_method'
 require 'storey/railtie' if defined?(Rails)
 require 'storey/exceptions'
 require 'storey/migrator'
 require 'storey/duplicator'
 require 'storey/hstore'
-require 'storey/dumper'
 require 'storey/ruby_dumper'
 require 'storey/sql_dumper'
+require 'storey/dumper'
 require 'storey/native_schema_matcher'
 require 'storey/suffixifier'
 require 'storey/unsuffixifier'
 require 'storey/resets_column_info'
+require 'storey/utils'
+require 'storey/builds_dump_command'
+require 'storey/builds_load_command'
+require 'storey/schema_name'
 
 module Storey
-  RESERVED_SCHEMAS = %w(hstore)
 
   mattr_accessor :suffix, :persistent_schemas
   mattr_writer :default_search_path
@@ -60,17 +64,12 @@ module Storey
   end
 
   def create(name, options={}, &block)
-    if name.blank?
-      fail ArgumentError, "Must pass in a valid schema name"
-    end
-
-    if RESERVED_SCHEMAS.include?(name) && !options[:force]
-      fail ArgumentError, "'#{name}' is a reserved schema name"
-    end
+    name = SchemaName.new(name)
+    name.validate_format!
+    name.validate_reserved! unless options[:force]
 
     if self.schemas.include?(name)
-      fail(Storey::SchemaExists,
-           %{The schema "#{name}" already exists.})
+      fail(Storey::SchemaExists, %{The schema "#{name}" already exists.})
     end
 
     if options[:load_database_structure].nil?
@@ -93,20 +92,8 @@ module Storey
 
   def create_plain_schema(schema_name)
     name = suffixify schema_name
-    command = %{"CREATE SCHEMA #{name}"}
-    switches = self.command_line_switches(command: command)
-    `psql #{switches}`
-  end
-
-  def command_line_switches(options={})
-    switches = {}
-    if self.database_config.has_key?(:host)
-      switches[:host] = self.database_config[:host]
-    end
-    switches[:dbname] = self.database_config[:database]
-    switches[:username] = self.database_config[:username]
-    switches = switches.merge(options)
-    switches.map {|k, v| "--#{k}=#{v}"}.join(' ')
+    command = "CREATE SCHEMA #{name}"
+    system psql_load_command(command: command)
   end
 
   def schemas(options={})
@@ -208,6 +195,10 @@ module Storey
     self.default_search_path == schema_name
   end
 
+  def db_command_line_switches_from(extra_config={})
+    Utils.db_command_line_switches_from(self.database_config, extra_config)
+  end
+
   protected
 
   def schema_migrations
@@ -246,6 +237,10 @@ module Storey
     ::ActiveRecord::Base.descendants.each do |descendant|
       descendant.reset_column_information
     end
+  end
+
+  def psql_load_command(options={})
+    BuildsLoadCommand.execute(self.database_config.merge(options))
   end
 
 end