+++ lib/storey/duplicator.rb
@@ -1,86 +1,84 @@
 module Storey
   class Duplicator
 
-    attr_accessor(:source_schema,
-                  :target_schema,
-                  :source_file,
-                  :target_file,
-                  :structure_only,
-                  :file_prefix,
-                  :dump_path,
-                  :source_dump_path,
-                  :target_dump_path)
-
     def initialize(from_schema, to_schema, options={})
       unless from_schema
         fail SchemaNotFound, "cannot duplicate from nil schema"
       end
 
-      self.dump_path = File.join Rails.root, 'tmp', 'schema_dumps'
-      self.source_dump_path = File.join self.dump_path, 'source'
-      self.target_dump_path = File.join self.dump_path, 'target'
-      self.structure_only = options[:structure_only] || false
+      @dump_dir = File.join(Rails.root, 'tmp', 'schema_dumps')
+      @source_dump_dir = File.join(@dump_dir, 'source')
+      @target_dump_dir = File.join(@dump_dir, 'target')
+      @structure_only = options[:structure_only] || false
 
-      self.source_schema = suffixify(from_schema)
-      self.target_schema = suffixify(to_schema)
-      self.file_prefix = "#{Time.now.to_i}_#{rand(100000)}"
-      self.source_file = File.join(
-        self.source_dump_path,
-        "#{self.file_prefix}_#{self.source_schema}.sql"
-      )
-      self.target_file = File.join(
-        self.target_dump_path,
-        "#{self.file_prefix}_#{self.target_schema}.sql"
-      )
+      @source_schema = suffixify(from_schema)
+      @target_schema = suffixify(to_schema)
+      @file_prefix = "#{Time.now.to_i}_#{rand(100000)}"
+      @source_file = File.join(@source_dump_dir,
+                               "#{@file_prefix}_#{@source_schema}.sql")
+      @target_file = File.join(@target_dump_dir,
+                               "#{@file_prefix}_#{@target_schema}.sql")
     end
 
     def perform!
       dump_schema
       replace_occurrences
       load_schema
+      clean_source
+      clean_target
     end
 
     private
+
+    def clean_source
+      FileUtils.rm(@source_file)
+    end
+
+    def clean_target
+      FileUtils.rm(@target_file)
+    end
 
     def dump_schema(options={})
       ENV['PGPASSWORD'] = Storey.database_config[:password]
       prepare_schema_dump_directories
 
-      unless Storey.database_config[:host].blank?
+      if Storey.database_config[:host].present?
         options[:host] ||= Storey.database_config[:host]
       end
-      options[:username] ||= Storey.database_config[:username]
-      options[:file]     ||= self.source_file
-      options[:schema]   ||= self.source_schema
 
-      switches = options.map { |k, v| "--#{k}=#{v}" }
-      switches << '--schema-only' if self.structure_only
-      switches = switches.join(" ")
+      arg_options = options.dup
+      arg_options[:username] ||= Storey.database_config[:username]
+      arg_options[:file]     ||= @source_file
+      arg_options[:schema]   ||= @source_schema
+      arg_options['schema-only'] = nil if @structure_only
+
+      switches = Utils.command_line_switches_from(arg_options)
 
       success = system("pg_dump #{switches} #{Storey.database_config[:database]}")
       unless success
-        raise StoreyError, "There seems to have been a problem dumping `#{self.source_schema}` to make a copy of it into `#{self.target_schema}`"
+        raise StoreyError, "There seems to have been a problem dumping `#{@source_schema}` to make a copy of it into `#{@target_schema}`"
       end
     end
 
     def prepare_schema_dump_directories
-      [self.source_dump_path, self.target_dump_path].each do |d|
+      [@source_dump_dir, @target_dump_dir].each do |d|
         FileUtils.mkdir_p(d)
       end
     end
 
     def load_schema(options={})
-      options[:file] ||= self.target_file
-      switches = Storey.command_line_switches(options)
+      options[:file] ||= @target_file
+      psql_options = Storey.database_config.merge(options)
 
       if duplicating_from_default?
         # Since we are copying the source schema and we're after structure only,
         # the dump_schema ended up creating a SQL file without the "CREATE SCHEMA" command
         # thus we have to create it manually
-        ::Storey.create_plain_schema self.target_schema
+        ::Storey.create_plain_schema @target_schema
       end
 
-      `psql #{switches}`
+      psql_load_command = BuildsLoadCommand.execute(psql_options)
+      system psql_load_command
 
       copy_source_schema_migrations
 
@@ -88,7 +86,7 @@ module Storey
     end
 
     def copy_source_schema_migrations
-      ::Storey.switch self.target_schema do
+      ::Storey.switch @target_schema do
         source_schema_migrations.each do |version|
           unless target_schema_migrations.include?(version)
             command = "INSERT INTO schema_migrations (version) VALUES ('#{version}');"
@@ -99,28 +97,28 @@ module Storey
     end
 
     def source_schema_migrations
-      ::Storey.switch(self.source_schema) do
+      ::Storey.switch(@source_schema) do
         ::ActiveRecord::Migrator.get_all_versions
       end
     end
 
     def target_schema_migrations
-      ::Storey.switch(self.target_schema) do
+      ::Storey.switch(@target_schema) do
         ::ActiveRecord::Migrator.get_all_versions
       end
     end
 
     def replace_occurrences
-      File.open(self.source_file, 'r') do |file|
+      File.open(@source_file, 'r') do |file|
         file.each_line do |line|
-          new_line = line.gsub(/#{self.source_schema}/, self.target_schema)
-          File.open(self.target_file, 'a') {|tf| tf.puts new_line}
+          new_line = line.gsub(/#{@source_schema}/, @target_schema)
+          File.open(@target_file, 'a') {|tf| tf.puts new_line}
         end
       end
     end
 
     def duplicating_from_default?
-      ::Storey.matches_default_search_path?(self.source_schema) && self.structure_only
+      ::Storey.matches_default_search_path?(@source_schema) && @structure_only
     end
 
     def suffixify(schema_name)